Linux compilation fix
Bench: 12012658

+++ src/timemanager.cpp
@@ -34,7 +34,7 @@ namespace Zagreus {
         }
 
         uint64_t movesToGo = params.movestogo ? params.movestogo : 40ULL - (bitboard.getPly() / 2);
-        movesToGo = std::max(movesToGo, 7ULL);
+        movesToGo = std::max((uint64_t) movesToGo, (uint64_t) 7ULL);
         uint64_t timeLeft = 0;
 
         if (movingColor == PieceColor::WHITE) {
@@ -46,7 +46,7 @@ namespace Zagreus {
         }
 
         timeLeft -= engine.getOption("Move Overhead").getIntValue();
-        timeLeft = std::max(timeLeft, 1ULL);
+        timeLeft = std::max((uint64_t) timeLeft, (uint64_t) 1ULL);
         uint64_t maxTime = timeLeft / 100 * 80;
         uint64_t timePerMove = timeLeft / movesToGo;
 
@@ -54,7 +54,7 @@ namespace Zagreus {
             timePerMove = maxTime;
         }
 
-        timePerMove = std::max(timePerMove, 1ULL);
+        timePerMove = std::max((uint64_t) timePerMove, (uint64_t) 1ULL);
 
         return std::chrono::high_resolution_clock::now() + std::chrono::milliseconds(timePerMove);
     }
